finish rename gcs -> storage
