+++ cmd/controller/main.go
@@ -21,8 +21,8 @@ import (
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 
 	"github.com/google/knative-gcp/pkg/reconciler/channel"
-	"github.com/google/knative-gcp/pkg/reconciler/gcs"
 	"github.com/google/knative-gcp/pkg/reconciler/pullsubscription"
+	"github.com/google/knative-gcp/pkg/reconciler/storage"
 	"github.com/google/knative-gcp/pkg/reconciler/topic"
 
 	"knative.dev/pkg/injection/sharedmain"
@@ -30,7 +30,7 @@ import (
 
 func main() {
 	sharedmain.Main("controller",
-		gcs.NewController,
+		storage.NewController,
 		pullsubscription.NewController,
 		topic.NewController,
 		channel.NewController,
